requirements.txt
asn1crypto==1.5.1
backports-zoneinfo==0.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.0.0
cryptography==45.0.5
filelock==3.16.1
idna==3.10
packaging==25.0
platformdirs==4.3.6
protobuf==5.29.5
pycparser==2.22
PyJWT==2.9.0
pyopenssl==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
six==1.17.0
snowflake-connector-python==3.14.0
snowflake-snowpark-python==1.30.0
sortedcontainers==2.4.0
tomlkit==0.13.3
typing-extensions==4.13.2
tzlocal==5.2
urllib3==1.26.20
